signature-core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
jinja2==3.1.4
beautifulsoup4==4.12.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
qrcode[pil]==7.4.2
Pillow==10.4.0
httpx==0.27.2
